Default untyped integer literals to be CGFloat ReactiveProperties.
Summary: Because we use CGFloat for nearly all of our operator types, `createProperty` should default to creating CGFloat properties when provided with integer literals. Clients can still create Int reactive properties by using the ReactiveProperty initializer directly.

Reviewers: O2 Material Motion, O4 Material Apple platform reviewers, #material_motion, chuga

Reviewed By: O4 Material Apple platform reviewers, chuga

Tags: #material_motion

Differential Revision: http://codereview.cc/D2904

+++ src/Timeline.swift
@@ -47,7 +47,7 @@ public final class Timeline {
 
    Only affects associated interactions if the timeline is paused.
    */
-  public let timeOffset = createProperty("Timeline.timeOffset", withInitialValue: CGFloat(0))
+  public let timeOffset = createProperty("Timeline.timeOffset", withInitialValue: 0)
 
   public let metadata = Metadata("Timeline")
 }
